fix: isolate worktree porcelain parser from electron imports
Agent-Logs-Url: https://github.com/Ron537/DPlex/sessions/dd8fa001-366e-4bff-b63d-625d9568ffd2
