+++ src/main/services/worktrees/gitWorktree.ts
@@ -13,6 +13,7 @@ import {
 } from '../gitService'
 import type { WorktreeInfo, WorktreeStatus } from './types'
 import { getEntriesForRepo } from './sidecar'
+import { parsePorcelain as parsePorcelainImpl, type RawWorktreeRecord } from './porcelain'
 
 /**
  * Parse `git worktree list --porcelain` output.
@@ -25,63 +26,6 @@ import { getEntriesForRepo } from './sidecar'
  *   [prunable <reason>]
  *   [locked <reason>]
  */
-interface RawWorktreeRecord {
-  path: string
-  head: string
-  branch: string | null
-  detached: boolean
-  bare: boolean
-  prunable: boolean
-}
-
-function parsePorcelainImpl(output: string): RawWorktreeRecord[] {
-  const records: RawWorktreeRecord[] = []
-  let cur: Partial<RawWorktreeRecord> | null = null
-
-  const flush = (): void => {
-    if (cur && cur.path) {
-      records.push({
-        path: cur.path,
-        head: cur.head ?? '',
-        branch: cur.branch ?? null,
-        detached: cur.detached ?? false,
-        bare: cur.bare ?? false,
-        prunable: cur.prunable ?? false
-      })
-    }
-    cur = null
-  }
-
-  for (const line of output.split('\n')) {
-    if (line.length === 0) {
-      flush()
-      continue
-    }
-    if (line.startsWith('worktree ')) {
-      flush()
-      cur = { path: line.slice('worktree '.length) }
-      continue
-    }
-    if (!cur) continue
-    if (line.startsWith('HEAD ')) {
-      cur.head = line.slice('HEAD '.length)
-    } else if (line.startsWith('branch ')) {
-      const ref = line.slice('branch '.length)
-      cur.branch = ref.startsWith('refs/heads/') ? ref.slice('refs/heads/'.length) : ref
-      cur.detached = false
-    } else if (line === 'detached') {
-      cur.branch = null
-      cur.detached = true
-    } else if (line === 'bare') {
-      cur.bare = true
-    } else if (line.startsWith('prunable')) {
-      cur.prunable = true
-    }
-  }
-  flush()
-  return records
-}
-
 /**
  * Exported for unit testing. Parses `git worktree list --porcelain` output into
  * structured records.
